logic update.
