+++ tests/benchmark/tpch/q20.py
@@ -20,6 +20,9 @@ if __name__ == "__main__":
     dataset_id, session, suffix = utils.get_tpch_configuration()
     current_path = Path(__file__).absolute()
 
+    if not session._strictly_ordered:
+        exit()
+
     utils.get_execution_time(
         vendored_tpch_q20.q, current_path, suffix, dataset_id, session
     )
